fix: 遊戲存活的玩家，手中的武器應該全部刷新(子彈)
