+++ scripts/modules/hotbar/Hotbar.ts
@@ -1,5 +1,7 @@
 import { ItemStack, Player } from "@minecraft/server";
 import { Glock17 } from "../weapon/actors/item/Glock17";
+import { ActorManager } from "../weapon/systems/ActorManager";
+import { ItemActor } from "../weapon/actors/Actor";
 
 export interface IHotbar {
     readonly items: Array<ItemStack|undefined>;
@@ -13,19 +15,20 @@ class Hotbar {
         this.items = new Array(9).fill(undefined);
     }
 
-    copy(target: Player) {
+    copyInventory(target: Player) {
         const container = target.getComponent('inventory')!.container;
         for (let index = 0; index < 9; index ++) {
             this.items[index] = container.getItem(index);
         }
     }
+
 }
 
 export class HotbarManager {
 
     static getPlayerHotbar(player: Player): IHotbar {
         const hotbar = new Hotbar();
-        hotbar.copy(player);
+        hotbar.copyInventory(player);
         return hotbar;
     }
 
@@ -35,6 +38,21 @@ export class HotbarManager {
             container.setItem(index, hotbar.items.at(index));
         }
     }
+
+    static resetItemActors(hotbar: IHotbar) {
+        for (let index = 0; index < 9; index ++) {
+            const item = hotbar.items[index];
+            if (item === undefined || !ActorManager.isActor(item)) continue;
+
+            const itemActor = ActorManager.getActor(item) as ItemActor;
+            const cloneActor = itemActor.clone();
+            hotbar.items[index] = cloneActor.item;
+
+            ActorManager.removeActor(itemActor.uuid);
+        }
+        return hotbar;
+    }
+
 }
 
 export class HotbarTemplate {
